Move theme switcher and provider to components folder
