+++ packages/titus-kitchen-sink/src/components/header-right/header-right.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import UserProfile from '../user-profile/user-profile'
-import { ThemeSwitcher } from '../../theme'
+import ThemeSwitcher from '../theme-switcher/theme-switcher'
 
 const HeaderRight = (props) => (
   <div {...props}>
