+++ packages/titus-kitchen-sink/src/components/theme-switcher/theme-switcher.js
@@ -3,8 +3,8 @@ import { Query } from 'react-apollo'
 import IconButton from '@material-ui/core/IconButton'
 import InvertColorsIcon from '@material-ui/icons/InvertColors'
 
-import { getThemeName } from '-!graphql-tag/loader!./queries.gql' // eslint-disable-line import/no-webpack-loader-syntax
-import { THEME } from './theme'
+import { getThemeName } from '-!graphql-tag/loader!../../theme/queries.gql' // eslint-disable-line import/no-webpack-loader-syntax
+import { THEME } from '../../theme'
 
 const ThemeSwitcher = () => (
   <Query query={getThemeName}>
