+++ packages/titus-kitchen-sink/src/theme/index.js
@@ -1,6 +1,1 @@
-import ThemeSwitcher from './theme-switcher'
-import ThemeProvider from './theme-provider'
-
 export * from './theme'
-
-export {ThemeSwitcher, ThemeProvider}
